fr-FR: Import the French translation
This is the original content of the « Debian Squeeze » book published
by Eyrolles that was used to create the « Debian Administrator's
Handbook ». This content is now released under the same terms as the
English book thanks to a successful crowdfunding campaign:
http://www.ulule.com/liberation-cahier-admin-debian/

All the translations comes from the original PO files used to
translate the French book into English, with the msgid/msgtxt strings
simply reversed. Some further work is needed to synchronize it
with the strings changes that happened since the release of the English
book.