Update epub build process
We're now using kindlegen when available as it provides a file
that renders good on modern readers supporting the KF8 format
while still staying compatible with the original mobipocket
format.

We also use media css to have a single CSS file. However we use
another XSL file to build the mobipocket version with different
options that generate non-modern HTML but that at least renders
correctly with kindlegen.