Ensure WeakReference is never null
In some extreme cases, the system will kill the whole app to free some memory. Since services are important, they might be rebooted first, in which case the static Permisision.mainActivity will be undefined. This commit sets it to a null WeakReference to make sure we never get a NullPointerException on it.