Removed the app sort options
Honestly it doesn't make any sense, we should have only use 'A to Z' all the time. There is no reason to offer this as an option, and the new fastscroll code would make maintenance for this a nightmare. We might as well just get rid of it now, esp. since A to Z is now the default option anyway for new installs.