Use contact ID rather than lookup key
Contact history was implemented using LOOKUP_KEY, as it seemed to be good practice in the Android world.
However, LOOKUP_KEY can and will change.

At this point, we might as well use the shorter contact id, that may also change but seems to change less frequently and is much shorter to store and manipulate.