Got it working!
Issue was caused by the use of setOnTouchListener(), which disables default animation.
You need to return false from onTouch() to get the default behavior, otherwise the touch event is 'consumed' and you won't see the animation.

You also need to mark the view as android:clickable, with the correct background type.

In the end, fix was easy, but hard to find.