Revamp long press detection
Widgets that do not capture clicks were not displaying the widget menu.
With this PR, all widgets should now properly display the context menu.

This also ensures that menu are displayed as PopupMenu, around the widget, allowing us to get rid of the currentlySelectedWidget attribute, and keeping the onOptionsItemSelected() function for MainActivity events only.