Disclaimer for Accessibility API
The Play Store notified me that each app making use of the accessibility
API needs to have multiple disclaimers: one in the store listing, and
another one when the feature is enabled.

The rules are complex:

-----------

App Status: Rejected

Your app has been rejected and wasn't published due to the policy
issue(s) listed below. If you submitted an update, the previous version
of your app is still available on Google Play.

Issue found: Missing description in Play Listing

We were unable to approve your app because the app description on the
play store does not reflect the use of AccessibilityService API.

Please update your app's long description so that it meets all the
policy requirements.

About the Accessibility API Policy

The Accessibility API cannot be used to:

 * Change user settings without their permission or prevent the ability
   for users to disable or uninstall any app or service unless
authorized by a parent or guardian through a parental control app or by
authorized administrators through enterprise management software.
 * Work around Android built-in privacy controls and notifications or
 * Change or leverage the user interface in a way that is deceptive or
   otherwise violates Play Developer Policies.
The accessibility API is not designed and cannot be requested for remote
call audio recording.

The use of the Accessibility API must be documented in the Play Store
listing.

Apps with a core functionality intended to directly support people with
disabilities are eligible to use the IsAccessibilityTool to
appropriately publicly designate themselves as an accessibility app.
Apps not eligible for IsAccessibilityTool may not use the flag and must
meet prominent disclosure and consent requirements as outlined in the
User Data policy as the accessibility related functionality is not
obvious to the user. Please refer to the AccessibilityService API help
center article for more information.

Apps must use limited, more narrowly scoped APIs and permissions in lieu
of the Accessibility API when possible to achieve the desired
functionality.

Additionally, follow these steps to bring your app into compliance: Read
more about AccessibilityService API Capability and Prominent Disclosure
best practices

Issue found: Insufficient prominent disclosure

We were unable to approve your app because the prominent disclosure does
not provide sufficient details on the use of AccessibilityService API.

The prominent disclosure should explain why the AccessibilityService API
is needed with details of relevant features in the app.

Please update your app so that the prominent disclosure meets all the
requirements as defined in the User Data Policy.

About the Accessibility API Policy

The Accessibility API cannot be used to:

 * Change user settings without their permission or prevent the ability
   for users to disable or uninstall any app or service unless
authorized by a parent or guardian through a parental control app or by
authorized administrators through enterprise management software.
 * Work around Android built-in privacy controls and notifications or
 * Change or leverage the user interface in a way that is deceptive or
   otherwise violates Play Developer Policies.
The accessibility API is not designed and cannot be requested for remote
call audio recording.

The use of the Accessibility API must be documented in the Play Store
listing.

Apps with a core functionality intended to directly support people with
disabilities are eligible to use the IsAccessibilityTool to
appropriately publicly designate themselves as an accessibility app.
Apps not eligible for IsAccessibilityTool may not use the flag and must
meet prominent disclosure and consent requirements as outlined in the
User Data policy as the accessibility related functionality is not
obvious to the user. Please refer to the AccessibilityService API help
center article for more information.

Apps must use limited, more narrowly scoped APIs and permissions in lieu
of the Accessibility API when possible to achieve the desired
functionality.