Do not hide keyboard when SearchEditText is inactive view
This fixes regression from #2280 which may cause the menu of search
result items to be cut off after long pressing the item. The problem is
that after the keyboard becomes hidden after a long press, the
resulting menu may then be moved down which can cause it to be cut
off. The solution is to only hide keyboard when SearchEditText is
the active view.