fix: cannot find some contacts which exist
Some contacts couldn't be opened, Contacts' app tells it doesn't exist although they do.

It seems that lookupId isn't (sadly) enough to find a contact in the app,
a contactId must be added in order for the CONTENT_URI to find the contact.

This format is known as the "complete" format, as described in Android Documentation.
https://developer.android.com/reference/android/provider/ContactsContract.Contacts#CONTENT_LOOKUP_URI